+++ src/media_centers/kodi/v16/MovieXmlWriterV16.cpp
@@ -131,17 +131,17 @@ QByteArray MovieXmlWriterV16::getMovieXml()
 
     KodiXml::removeChildNodes(doc, "actor");
 
-    for (const Actor& actor : m_movie.actors()) {
+    for (const Actor* actor : m_movie.actors()) {
         QDomElement elem = doc.createElement("actor");
         QDomElement elemName = doc.createElement("name");
         QDomElement elemRole = doc.createElement("role");
-        elemName.appendChild(doc.createTextNode(actor.name));
-        elemRole.appendChild(doc.createTextNode(actor.role));
+        elemName.appendChild(doc.createTextNode(actor->name));
+        elemRole.appendChild(doc.createTextNode(actor->role));
         elem.appendChild(elemName);
         elem.appendChild(elemRole);
-        if (!actor.thumb.isEmpty() && Settings::instance()->advanced()->writeThumbUrlsToNfo()) {
+        if (!actor->thumb.isEmpty() && Settings::instance()->advanced()->writeThumbUrlsToNfo()) {
             QDomElement elemThumb = doc.createElement("thumb");
-            elemThumb.appendChild(doc.createTextNode(actor.thumb));
+            elemThumb.appendChild(doc.createTextNode(actor->thumb));
             elem.appendChild(elemThumb);
         }
         KodiXml::appendXmlNode(doc, elem);